requirements.txt
aioboto3==13.4.0
boto3==1.36.0
fastapi==0.115.7
pydantic==2.10.5
python_jose==3.3.0
redis==5.2.1
SQLAlchemy==2.0.36
starlette==0.45.2
uvicorn==0.34.0
pymysql==1.1.1
prometheus-fastapi-instrumentator==7.0.2
python-dotenv==1.0.1
boto3==1.36.0